bookworm updates

+++ lib/screens/bookworm/folders/book_options.dart
@@ -1,6 +1,15 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:mp3_music_converter/screens/bookworm/model/model.dart';
+import 'package:mp3_music_converter/screens/bookworm/provider/bookworm_provider.dart';
 import 'package:mp3_music_converter/screens/bookworm/services/book_services.dart';
+import 'package:http/http.dart' as http;
+import 'package:mp3_music_converter/utils/helper/helper.dart';
+import 'package:mp3_music_converter/utils/helper/instances.dart';
+import 'package:mp3_music_converter/widgets/progress_indicator.dart';
+import 'package:provider/provider.dart';
 
 showBookOptions(BuildContext context, Book book) {
   return showDialog(
@@ -56,8 +65,47 @@ class ShowConfirmRemoveBookOption extends StatefulWidget {
 class _ShowConfirmRemoveBookOptionState
     extends State<ShowConfirmRemoveBookOption> {
   removeBook() async {
-    await BookwormServices().deleteBook(widget.book);
-    Navigator.pop(context);
+    String url = 'https://youtubeaudio.com/api/book/deletebook';
+    String token = await preferencesHelper.getStringValues(key: 'token');
+    CustomProgressIndicator _progressIndicator =
+        CustomProgressIndicator(context);
+
+    try {
+      _progressIndicator.show();
+      final response = await http.post(
+        url,
+        headers: {'Content-Type': 'application/json'},
+        body: jsonEncode(
+          {'token': token, 'id': widget.book.id},
+        ),
+      );
+      _progressIndicator.dismiss();
+      print(jsonDecode(response.body));
+      if (response.statusCode == 200) {
+        final data = jsonDecode(response.body);
+        if (data['message'].toString().toLowerCase().contains('delete')) {
+          Navigator.pop(context);
+          await BookwormServices().deleteBook(widget.book);
+          showToast(context,
+              message: 'Book deleted', backgroundColor: Colors.green);
+          Provider.of<BookwormProvider>(context, listen: false)
+              .getFolderContents(
+                  Provider.of<BookwormProvider>(context, listen: false)
+                      .currentFolder
+                      .name);
+          print('happened');
+        }
+      } else {
+        if (_progressIndicator.isShowing()) await _progressIndicator.dismiss();
+        final data = jsonDecode(response.body);
+        Navigator.pop(context);
+        showToast(context,
+            message: data['message'], backgroundColor: Colors.red);
+      }
+    } catch (e) {
+      if (_progressIndicator.isShowing()) await _progressIndicator.dismiss();
+      Navigator.pop(context);
+    }
   }
 
   @override
